move UnitParse functions to their own file
