+++ model/Clinical/CMDecisionTree.cpp
@@ -24,7 +24,7 @@
 #include "PkPd/LSTMTreatments.h"
 #include "util/random.h"
 #include "util/errors.h"
-#include "util/timeConversions.h"
+#include "util/UnitParse.h"
 #include "interventions/Interfaces.hpp"
 
 #include <limits>
