+++ model/Clinical/ClinicalModel.cpp
@@ -28,7 +28,7 @@
 #include "util/ModelOptions.h"
 #include "util/CommandLine.h"
 #include "util/errors.h"
-#include "util/timeConversions.h"
+#include "util/UnitParse.h"
 #include "schema/scenario.h"
 
 namespace OM { namespace Clinical {
