+++ model/WithinHost/WHFalciparum.cpp
@@ -35,7 +35,7 @@
 #include "util/errors.h"
 #include "util/StreamValidator.h"
 #include "util/checkpoint_containers.h"
-#include "util/timeConversions.h"
+#include "util/UnitParse.h"
 #include "schema/scenario.h"
 
 #include <cmath>
