+++ model/WithinHost/WHVivax.cpp
@@ -27,7 +27,7 @@
 #include "util/random.h"
 #include "util/errors.h"
 #include "util/checkpoint_containers.h"
-#include "util/timeConversions.h"
+#include "util/UnitParse.h"
 #include "util/CommandLine.h"
 #include "util/sampler.h"
 #include <schema/scenario.h>
