+++ model/util/DecayFunction.cpp
@@ -22,7 +22,7 @@
 #include "util/errors.h"
 #include "util/StreamValidator.h"
 #include "util/random.h"
-#include "util/timeConversions.h"
+#include "util/UnitParse.h"
 
 #include <cmath>
 #include <stdexcept>
